Revert to simpler update site building
Eclipse Marketplace seems to ignore `Solution Version` as the version
number of what is trying to be installed and instead goes for the latest
feature version available. This means even though the update site might
be hosting multiple versions, registered correctly in the content.jar,
we can't control version release via the Eclipse Marketplace, it will
always install the latest available feature.